[workspace] Split up pybind11 vs mkdoc
