+++ bindings/pydrake/BUILD.bazel
@@ -16,11 +16,14 @@ load(
     "//tools/skylark:pybind.bzl",
     "drake_pybind_cc_googletest",
     "drake_pybind_library",
-    "generate_pybind_documentation_header",
     "get_drake_py_installs",
     "get_pybind_package_info",
 )
 load("//tools/workspace:generate_file.bzl", "generate_file")
+load(
+    "//tools/workspace/mkdoc_internal:defs.bzl",
+    "generate_pybind_documentation_header",
+)
 
 package(default_visibility = [
     "//bindings:__subpackages__",
