+++ tools/workspace/mkdoc_internal/README.md
@@ -1,0 +1,5 @@
+This directory contains Drake's heavily customized fork of
+ https://github.com/pybind/pybind11_mkdoc
+that was actually forked from
+ https://github.com/pybind/pybind11
+before the pybind11_mkdoc separate repository even existed.
