+++ tools/workspace/mkdoc_internal/mkdoc.py
@@ -49,8 +49,8 @@ import sys
 from clang import cindex
 from clang.cindex import AccessSpecifier, CursorKind, TypeKind
 
-from tools.workspace.pybind11.libclang_setup import add_library_paths
-from tools.workspace.pybind11.mkdoc_comment import process_comment
+from tools.workspace.mkdoc_internal.libclang_setup import add_library_paths
+from tools.workspace.mkdoc_internal.mkdoc_comment import process_comment
 
 
 CLASS_KINDS = [
