+++ tools/workspace/mkdoc_internal/test/sample_header_documentation.expected.h
@@ -12,13 +12,13 @@
 #pragma GCC diagnostic ignored "-Wunused-variable"
 #endif
 
-// #include "drake/tools/workspace/pybind11/test/sample_header.h"
+// #include "drake/tools/workspace/mkdoc_internal/test/sample_header.h"
 
 // Symbol: sample_header_doc
 constexpr struct /* sample_header_doc */ {
   // Symbol: RootLevelSymbol
   struct /* RootLevelSymbol */ {
-    // Source: drake/tools/workspace/pybind11/test/sample_header.h:31
+    // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:31
     const char* doc =
 R"""(Root-level symbol. Magna fermentum iaculis eu non diam phasellus
 vestibulum.)""";
@@ -27,7 +27,7 @@ vestibulum.)""";
   struct /* drake */ {
     // Symbol: drake::MidLevelSymbol
     struct /* MidLevelSymbol */ {
-      // Source: drake/tools/workspace/pybind11/test/sample_header.h:51
+      // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:51
       const char* doc =
 R"""(1. Begin first ordered list element. Rutrum quisque non tellus orci ac
 auctor. End first ordered list element. 2. Begin second ordered list
@@ -44,12 +44,12 @@ vivamus at augue eget arcu dictum varius.)""";
     struct /* mkdoc_test */ {
       // Symbol: drake::mkdoc_test::AnonymousConstant
       struct /* AnonymousConstant */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:259
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:259
         const char* doc = R"""(Anonymous enum's constant.)""";
       } AnonymousConstant;
       // Symbol: drake::mkdoc_test::Class
       struct /* Class */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:75
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:75
         const char* doc =
 R"""(* Begin first unordered list element. Volutpat blandit aliquam etiam erat
   velit scelerisque. End first unordered list element.
@@ -60,41 +60,41 @@ R"""(* Begin first unordered list element. Volutpat blandit aliquam etiam erat
   sagittis purus sit. End third unordered list element.)""";
         // Symbol: drake::mkdoc_test::Class::Class
         struct /* ctor */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:97
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:97
           const char* doc_0args = R"""(Custom constructor 1.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:104
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:104
           const char* doc_1args =
 R"""(Custom constructor 2. Ut tristique et egestas quis ipsum suspendisse
 ultrices gravida. Suscipit tellus mauris a diam. Maecenas accumsan
 lacus vel facilisis volutpat est.
 
 Ut consequat semper viverra nam libero.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:112
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:112
           const char* doc_2args =
 R"""(Custom constructor 3. Integer quis auctor elit sed vulputate mi sit.)""";
         } ctor;
         // Symbol: drake::mkdoc_test::Class::Nested
         struct /* Nested */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:175
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:175
           const char* doc =
 R"""(Protected nested class. Sed turpis tincidunt id aliquet. Egestas sed
 sed risus pretium.)""";
         } Nested;
         // Symbol: drake::mkdoc_test::Class::ProtectedMethod
         struct /* ProtectedMethod */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:169
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:169
           const char* doc =
 R"""(Protected method. Nibh sed pulvinar proin gravida hendrerit. Orci
 phasellus egestas tellus rutrum tellus pellentesque eu.)""";
         } ProtectedMethod;
         // Symbol: drake::mkdoc_test::Class::PublicMethod
         struct /* PublicMethod */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:117
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:117
           const char* doc = R"""()""";
         } PublicMethod;
         // Symbol: drake::mkdoc_test::Class::PublicStatic
         struct /* PublicStatic */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:128
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:128
           const char* doc =
 R"""(Sed faucibus turpis in eu mi bibendum neque egestas.
 
@@ -109,63 +109,63 @@ Postcondition:
         } PublicStatic;
         // Symbol: drake::mkdoc_test::Class::PublicTemplateMethod
         struct /* PublicTemplateMethod */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:121
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:121
           const char* doc = R"""()""";
         } PublicTemplateMethod;
         // Symbol: drake::mkdoc_test::Class::TypedefAlias
         struct /* TypedefAlias */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:89
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:89
           const char* doc =
 R"""(Typedef alias. Risus nec feugiat in fermentum posuere urna nec
 tincidunt praesent.)""";
         } TypedefAlias;
         // Symbol: drake::mkdoc_test::Class::UsingAlias
         struct /* UsingAlias */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:84
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:84
           const char* doc =
 R"""(Using alias. Sit amet nisl purus in mollis nunc sed id semper.)""";
         } UsingAlias;
         // Symbol: drake::mkdoc_test::Class::do_stuff
         struct /* do_stuff */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:159
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:159
           const char* doc_stuff_1 = R"""(Docstring 1.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:164
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:164
           const char* doc_stuff_2 = R"""(Docstring 2.)""";
         } do_stuff;
         // Symbol: drake::mkdoc_test::Class::get_foo
         struct /* get_foo */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:152
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:152
           const char* doc_0args_nonconst = R"""(Overloaded only by its const-ness.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:155
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:155
           const char* doc_0args_const = R"""(The const one.)""";
         } get_foo;
         // Symbol: drake::mkdoc_test::Class::overloaded_method
         struct /* overloaded_method */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:131
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:131
           const char* doc_1args_alpha = R"""(This one takes an int.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:134
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:134
           const char* doc_1args_bravo = R"""(This one takes a double.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:137
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:137
           const char* doc_2args_charlie_delta = R"""(This one takes an int and a double.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:140
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:140
           const char* doc_2args_double_int = R"""(This one takes the road less traveled.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:143
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:143
           const char* doc_1args_conststdstring = R"""(This one takes a non-primitive type.)""";
         } overloaded_method;
         // Symbol: drake::mkdoc_test::Class::overloaded_with_same_doc
         struct /* overloaded_with_same_doc */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:146
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:146
           const char* doc = R"""(Different overload with same doc.)""";
         } overloaded_with_same_doc;
         // Symbol: drake::mkdoc_test::Class::protected_member_
         struct /* protected_member_ */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:177
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:177
           const char* doc = R"""()""";
         } protected_member_;
       } Class;
       // Symbol: drake::mkdoc_test::DrakeDeprecatedClass
       struct /* DrakeDeprecatedClass */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:264
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:264
         const char* doc_deprecated =
 R"""(I am measurably old. (Deprecated.)
 
@@ -174,7 +174,7 @@ Deprecated:
     after 2038-01-19.)""";
         // Symbol: drake::mkdoc_test::DrakeDeprecatedClass::a
         struct /* a */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:277
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:277
           const char* doc_deprecated =
 R"""((Deprecated.)
 
@@ -184,14 +184,14 @@ Deprecated:
         } a;
         // Symbol: drake::mkdoc_test::DrakeDeprecatedClass::f
         struct /* f */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:269
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:269
           const char* doc_deprecated_1args =
 R"""((Deprecated.)
 
 Deprecated:
     f() is slow; use g() instead. Also, I like hats. This will be
     removed from Drake on or after 2038-01-19.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:273
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:273
           const char* doc_deprecated_0args =
 R"""((Deprecated.)
 
@@ -202,7 +202,7 @@ Deprecated:
       } DrakeDeprecatedClass;
       // Symbol: drake::mkdoc_test::DrakeDeprecatedTemplateClass
       struct /* DrakeDeprecatedTemplateClass */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:283
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:283
         const char* doc_deprecated =
 R"""(I am symbolically old. (Deprecated.)
 
@@ -212,32 +212,32 @@ Deprecated:
       } DrakeDeprecatedTemplateClass;
       // Symbol: drake::mkdoc_test::Enum
       struct /* Enum */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:245
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:245
         const char* doc =
 R"""(Enumeration. Feugiat scelerisque varius morbi enim. Facilisis leo vel
 fringilla est ullamcorper eget nulla facilisi.)""";
         // Symbol: drake::mkdoc_test::Enum::EnumConstant
         struct /* EnumConstant */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:247
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:247
           const char* doc = R"""(Enumeration constant.)""";
         } EnumConstant;
       } Enum;
       // Symbol: drake::mkdoc_test::EnumClass
       struct /* EnumClass */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:252
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:252
         const char* doc =
 R"""(Enumeration class. Malesuada fames ac turpis egestas integer eget
 aliquet nibh praesent.)""";
         // Symbol: drake::mkdoc_test::EnumClass::EnumClassConstant
         struct /* EnumClassConstant */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:253
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:253
           const char* doc =
 R"""(Enumeration class constant. Vestibulum mattis.)""";
         } EnumClassConstant;
       } EnumClass;
       // Symbol: drake::mkdoc_test::Struct
       struct /* Struct */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:192
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:192
         const char* doc_deprecated =
 R"""(Struct. Sed elementum tempus egestas sed sed risus pretium. Vel
 pharetra vel turpis nunc.
@@ -247,20 +247,20 @@ Deprecated:
     tincidunt lobortis. End deprecated.)""";
         // Symbol: drake::mkdoc_test::Struct::Serialize
         struct /* Serialize */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:195
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:195
           const char* doc =
 R"""(See implementing_serialize "Implementing Serialize".)""";
         } Serialize;
         // Symbol: drake::mkdoc_test::Struct::field_1
         struct /* field_1 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:198
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:198
           const char* doc =
 R"""(Field 1. Sit amet cursus sit amet dictum sit amet. Id leo in vitae
 turpis massa sed elementum tempus.)""";
         } field_1;
         // Symbol: drake::mkdoc_test::Struct::field_2
         struct /* field_2 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:201
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:201
           const char* doc =
 R"""(Field 2. Consectetur libero id faucibus nisl tincidunt eget nullam non
 nisi.)""";
@@ -274,32 +274,32 @@ nisi.)""";
       } Struct;
       // Symbol: drake::mkdoc_test::TemplateClass
       struct /* TemplateClass */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:207
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:207
         const char* doc_was_unable_to_choose_unambiguous_names =
 R"""(Template class. Mauris pharetra et ultrices neque ornare aenean
 euismod elementum.)""";
         // Symbol: drake::mkdoc_test::TemplateClass::TemplateClass<T>
         struct /* ctor */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:213
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:213
           const char* doc_0args =
 R"""(Default constructor. Condimentum mattis pellentesque id nibh tortor
 id. Nisl rhoncus mattis rhoncus urna neque.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:216
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:216
           const char* doc_1args = R"""(Single argument int constructor.)""";
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:220
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:220
           const char* doc_copyconvert = R"""(Scalar-converting copy constructor.)""";
         } ctor;
       } TemplateClass;
       // Symbol: drake::mkdoc_test::func
       struct /* func */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:59
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:59
         const char* doc_0args =
 R"""(Function. Mi sit amet mauris commodo quis.)""";
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:62
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:62
         const char* doc_1args_param =
 R"""(Function, overload 1. Velit ut tortor pretium viverra suspendisse
 potenti nullam ac tortor.)""";
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:66
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:66
         const char* doc_1args_T =
 R"""(Function, template overload. Pellentesque diam volutpat commodo sed
 egestas egestas fringilla phasellus faucibus.)""";
@@ -310,7 +310,7 @@ egestas egestas fringilla phasellus faucibus.)""";
   struct /* namespace_1 */ {
     // Symbol: namespace_1::MySimpleSystem
     struct /* MySimpleSystem */ {
-      // Source: drake/tools/workspace/pybind11/test/sample_header.h:378
+      // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:378
       const char* doc =
 R"""(My simple system.
 
@@ -326,38 +326,38 @@ R"""(My simple system.
     struct /* namespace_2 */ {
       // Symbol: namespace_1::namespace_2::DummyClass
       struct /* DummyClass */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:329
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:329
         const char* doc = R"""()""";
         // Symbol: namespace_1::namespace_2::DummyClass::Details
         struct /* Details */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:334
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:334
           const char* doc =
 R"""(Ligula. Nunc turpis. Mauris vitae sapien. Nunc.)""";
         } Details;
         // Symbol: namespace_1::namespace_2::DummyClass::DummyClass
         struct /* ctor */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:331
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:331
           const char* doc = R"""()""";
         } ctor;
         // Symbol: namespace_1::namespace_2::DummyClass::static_function_1
         struct /* static_function_1 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:342
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:342
           const char* doc = R"""()""";
         } static_function_1;
         // Symbol: namespace_1::namespace_2::DummyClass::static_function_2
         struct /* static_function_2 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:343
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:343
           const char* doc = R"""()""";
         } static_function_2;
         // Symbol: namespace_1::namespace_2::DummyClass::struct_2
         struct /* struct_2 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:341
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:341
           const char* doc = R"""()""";
         } struct_2;
       } DummyClass;
       // Symbol: namespace_1::namespace_2::Struct1
       struct /* Struct1 */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:304
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:304
         const char* doc =
 R"""(Quam odio at est.
 
@@ -367,74 +367,74 @@ Nam quis tellus. Cras elit nisi, ornare a, condimentum vitae, rutrum
 sit amet, tellus. Maecenas)""";
         // Symbol: namespace_1::namespace_2::Struct1::var_1
         struct /* var_1 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:306
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:306
           const char* doc =
 R"""(Et, ornare sagittis, tellus. Fusce felis.)""";
         } var_1;
         // Symbol: namespace_1::namespace_2::Struct1::var_2
         struct /* var_2 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:308
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:308
           const char* doc = R"""(Nulla a augue. Pellentesque sed est.)""";
         } var_2;
         // Symbol: namespace_1::namespace_2::Struct1::var_3
         struct /* var_3 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:310
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:310
           const char* doc = R"""(Imperdiet tristique, interdum a, dolor.)""";
         } var_3;
         // Symbol: namespace_1::namespace_2::Struct1::var_4
         struct /* var_4 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:312
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:312
           const char* doc =
 R"""(Tempor lobortis turpis. Sed tellus velit, ullamcorper.)""";
         } var_4;
         // Symbol: namespace_1::namespace_2::Struct1::var_5
         struct /* var_5 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:314
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:314
           const char* doc =
 R"""(Id, rutrum auctor, ullamcorper sed, orci. In.)""";
         } var_5;
         // Symbol: namespace_1::namespace_2::Struct1::var_6
         struct /* var_6 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:316
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:316
           const char* doc =
 R"""(Fames ac turpis egestas. Sed vitae eros. Nulla.)""";
         } var_6;
         // Symbol: namespace_1::namespace_2::Struct1::var_7
         struct /* var_7 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:318
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:318
           const char* doc =
 R"""(Condimentum. Donec arcu quam, dictum accumsan, convallis.)""";
         } var_7;
         // Symbol: namespace_1::namespace_2::Struct1::var_8
         struct /* var_8 */ {
-          // Source: drake/tools/workspace/pybind11/test/sample_header.h:320
+          // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:320
           const char* doc =
 R"""(Volutpat. Donec non tortor. Vivamus posuere nisi mollis.)""";
         } var_8;
       } Struct1;
       // Symbol: namespace_1::namespace_2::Struct2
       struct /* Struct2 */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:323
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:323
         const char* doc = R"""()""";
       } Struct2;
       // Symbol: namespace_1::namespace_2::Struct3
       struct /* Struct3 */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:324
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:324
         const char* doc = R"""()""";
       } Struct3;
       // Symbol: namespace_1::namespace_2::Struct4
       struct /* Struct4 */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:325
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:325
         const char* doc = R"""()""";
       } Struct4;
       // Symbol: namespace_1::namespace_2::Struct5
       struct /* Struct5 */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:326
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:326
         const char* doc = R"""()""";
       } Struct5;
       // Symbol: namespace_1::namespace_2::Struct6
       struct /* Struct6 */ {
-        // Source: drake/tools/workspace/pybind11/test/sample_header.h:327
+        // Source: drake/tools/workspace/mkdoc_internal/test/sample_header.h:327
         const char* doc = R"""()""";
       } Struct6;
     } namespace_2;
